Added Version Key While Featching File From Git
The reason behind to do so is, when we pull code using raw.githubusercontent.com It sometmes return cached data insted of new commit from the github. So by adding a Version/v key wile featching file content and setting it's value to time() to always have a unique value for version key, returns the fresh featched data insted of cached.